requirements.txt
fastapi
uvicorn
psycopg2-binary
passlib
python-jose